Future: Add classes Future and Promise to allow asynchronous programming.
Both the class Future and Promise hold a shared state which controls
whether the promise has been fulfilled, and the object can be
retrieved. The shared state is stored out of class in the heap,
and will be deallocated when the last copy of the Future instances
deallocates.

The Promise is used to provide the shared state of the object the Future
awaits, and will ve valid as long as at least one instance of Future
exists.