Iterator: simplify iterators
No more BidirectionalIterator, ForwardIterator, Iterator.
Just Iterator, which can go forward and back.
The default iterator container has a reference to the container it is iterating over,
used to check range and item access.
Added is_end() to iterators, which allows to check for end of iterated collection.
Added index() to iterators, which allows to get the current index of the iterator.