Add GenericLexer
GenericLexer is a general-purpose rule-based lexer than can be used to tokenize
any input string given a correct set of rules. Construct the class with a list
of the lexing rules. The method tokenize(String const&) tokenizes the string argument
with the rules specified in the class constructor, and returns a vector of tokens, which can
then be parsed or processed.

For implementation details read GenericLexer.h or the generic_lexer test under the tests directory
in the source tree.